pyspark/wikimedia_streaming/requirements.txt
requests==2.31.0
requests-sse==0.3.1
google-cloud-pubsub==2.20.0